+++ src/BinomialTreeOption.py
@@ -1,7 +1,10 @@
 # Price an American option by the binomial tree model
-from StockOption import StockOption
 import math
+
 import numpy as np
+
+from src.StockOption import StockOption
+
 
 class BinomialTreeOption(StockOption):
     def _setup_parameters_(self):
@@ -10,7 +13,7 @@ class BinomialTreeOption(StockOption):
         self.d = 1 - self.pd # Expected value in the down state
         self.qu = (math.exp((self.r-self.div)*self.dt) -self.d) / (self.u-self.d)
         self.qd = 1-self.qu
-        
+
     def _initialize_stock_price_tree_(self):
          # Initialize a 2D tree at T=0
         self.STs = [np.array([self.S0])]
@@ -19,36 +22,36 @@ class BinomialTreeOption(StockOption):
             prev_branches = self.STs[-1]
             st = np.concatenate((prev_branches*self.u,[prev_branches[-1]*self.d]))
             self.STs.append(st) # Add nodes at each time step
-        
+
     def _initialize_payoffs_tree_(self):
         # Get payoffs when the option expires
         return np.maximum(0, (self.STs[self.N]-self.K) if self.is_call
         else(self.K-self.STs[self.N]))
-    
+
     def __check_early_exercise__(self, payoffs, node):
         early_ex_payoff = (self.STs[node] - self.K) if self.is_call else (self.K - self.STs[node])
-    
+
         return np.maximum(payoffs, early_ex_payoff)
-    
+
     def _traverse_tree_(self, payoffs):
         for i in reversed(range(self.N)):
-            # The payoffs from NOT exercising the option 
+            # The payoffs from NOT exercising the option
             payoffs = (payoffs[:-1] * self.qu + payoffs[1:] * self.qd) * self.df
 
             # The payoffs from exercising, for American options
             if not self.is_european:
                 payoffs = self.__check_early_exercise__(payoffs,i)
-        
+
         return payoffs
-    
+
     def __begin_tree_traversal__(self):
         payoffs = self._initialize_payoffs_tree_()
         return self._traverse_tree_(payoffs)
-    
+
     def price(self):
-        #The pricing implementation 
+        #The pricing implementation
         self._setup_parameters_()
         self._initialize_stock_price_tree_()
         payoffs = self.__begin_tree_traversal__()
-        
+
         return payoffs[0] # Option value converges to first node
